requirements.txt
Pillow
zarinpal-py-sdk
zeep
aiohttp
django-jazzmin

django-ckeditor